+++ double-map.go
@@ -61,12 +61,12 @@ func (s *DoubleMap) CountForSecondKey(key string) (int, error) {
 	return o.(int), nil
 }
 
-func (s *DoubleMap) Size(firstKey string, secondKey string) (int, error) {
+func (s *DoubleMap) Size(firstKey string, secondKey string) (int64, error) {
 	o, err := s.Client().SQLQueryFirst("select coalesce(length(value), 0) from $table$ where first_key=? and second_key=?;", IntScanner, firstKey, secondKey)
 	if err != nil {
 		return 0, err
 	}
-	return o.(int), nil
+	return o.(int64), nil
 }
 
 func (s *DoubleMap) TotalSize() (int64, error) {
